services/teamservice/requirements.txt
asgiref==3.11.0
Django==6.0
sqlparse==0.5.5
psycopg2-binary==2.9.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
